Move new backend functionality into own folder
