+++ server/endpoints/SearchBar.test.ts
@@ -3,8 +3,8 @@ import { MongoMemoryServer } from 'mongodb-memory-server';
 import express from "express";
 
 import axios from 'axios';
-import { configure } from "./endpoints";
-import { Classes, Students, Subjects } from './dbDefs';
+import { configure } from "../endpoints";
+import { Classes, Students, Subjects } from "../dbDefs";
 
 let mongoServer: MongoMemoryServer;
 let serverCloseHandle;
